feat: extract registry implementation
