+++ crates/tinymist-package/Cargo.toml
@@ -1,0 +1,49 @@
+[package]
+name = "tinymist-package"
+description = "Tinymist package support for Typst."
+categories = ["compilers"]
+keywords = ["api", "language", "typst"]
+authors.workspace = true
+version.workspace = true
+license.workspace = true
+edition.workspace = true
+homepage.workspace = true
+repository.workspace = true
+rust-version.workspace = true
+
+[dependencies]
+
+base64.workspace = true
+comemo.workspace = true
+ecow.workspace = true
+flate2.workspace = true
+log.workspace = true
+parking_lot.workspace = true
+rayon.workspace = true
+serde.workspace = true
+serde_json.workspace = true
+tar.workspace = true
+typst.workspace = true
+tinymist-std.workspace = true
+
+reqwest = { workspace = true, optional = true }
+dirs = { workspace = true, optional = true }
+
+js-sys = { workspace = true, optional = true }
+wasm-bindgen = { workspace = true, optional = true }
+web-sys = { workspace = true, optional = true, features = ["console"] }
+
+[dev-dependencies]
+insta.workspace = true
+
+[features]
+
+default = []
+http-registry = ["reqwest", "dirs"]
+web = ["wasm-bindgen", "web-sys", "js-sys"]
+browser = ["web"]
+system = ["http-registry"]
+
+
+[lints]
+workspace = true
