+++ crates/tinymist-package/src/lib.rs
@@ -1,0 +1,4 @@
+//! Package Implementation for Typst.
+
+pub mod registry;
+pub use registry::{PackageError, PackageRegistry, PackageSpec};
