+++ crates/tinymist-package/src/registry/http.rs
@@ -1,4 +1,4 @@
-//! Https registry for tinymist.
+//! Http registry for tinymist.
 
 use std::path::Path;
 use std::sync::{Arc, OnceLock};
@@ -10,7 +10,7 @@ use tinymist_std::ImmutPath;
 use typst::diag::{eco_format, EcoString, PackageResult, StrResult};
 use typst::syntax::package::{PackageVersion, VersionlessPackageSpec};
 
-use crate::package::{DummyNotifier, Notifier, PackageError, PackageRegistry, PackageSpec};
+use super::{DummyNotifier, Notifier, PackageError, PackageRegistry, PackageSpec};
 
 /// The http package registry for typst.ts.
 pub struct HttpRegistry {
