+++ crates/tinymist-world/Cargo.toml
@@ -41,6 +41,7 @@ serde-wasm-bindgen = { workspace = true, optional = true }
 sha2.workspace = true
 strum.workspace = true
 tar.workspace = true
+tinymist-package.workspace = true
 tinymist-std.workspace = true
 tinymist-vfs.workspace = true
 typst.workspace = true
@@ -62,8 +63,9 @@ web = [
     "serde-wasm-bindgen",
     "tinymist-std/web",
     "tinymist-vfs/web",
+    "tinymist-package/web",
 ]
-browser = ["tinymist-vfs/browser", "web"]
+browser = ["web", "tinymist-vfs/browser", "tinymist-package/browser"]
 system = [
     "dirs",
     "fontdb",
@@ -71,6 +73,7 @@ system = [
     "http-registry",
     "tinymist-std/system",
     "tinymist-vfs/system",
+    "tinymist-package/system",
 ]
 
 [lints]
