+++ crates/tinymist-world/src/package/mod.rs
@@ -1,40 +1,9 @@
-impl Notifier for DummyNotifier {}
-use std::num::NonZeroUsize;
-use std::{path::Path, sync::Arc};
+use std::sync::Arc;
 
-use ecow::EcoString;
 use tinymist_std::ImmutPath;
 use typst::diag::FileResult;
-pub use typst::diag::PackageError;
-pub use typst::syntax::package::PackageSpec;
 
-pub mod dummy;
-
-#[cfg(feature = "browser")]
-pub mod browser;
-
-#[cfg(feature = "http-registry")]
-pub mod http;
-
-pub trait PackageRegistry {
-    fn reset(&mut self) {}
-
-    fn revision(&self) -> Option<NonZeroUsize> {
-        None
-    }
-
-    fn resolve(&self, spec: &PackageSpec) -> Result<Arc<Path>, PackageError>;
-
-    /// A list of all available packages and optionally descriptions for them.
-    ///
-    /// This function is optional to implement. It enhances the user experience
-    /// by enabling autocompletion for packages. Details about packages from the
-    /// `@preview` namespace are available from
-    /// `https://packages.typst.org/preview/index.json`.
-    fn packages(&self) -> &[(PackageSpec, Option<EcoString>)] {
-        &[]
-    }
-}
+pub use tinymist_package::*;
 
 pub struct RegistryPathMapper<T> {
     pub registry: Arc<T>,
@@ -51,10 +20,3 @@ impl<T: PackageRegistry> tinymist_vfs::RootResolver for RegistryPathMapper<T> {
         Ok(self.registry.resolve(pkg)?)
     }
 }
-
-pub trait Notifier {
-    fn downloading(&self, _spec: &PackageSpec) {}
-}
-
-#[derive(Debug, Default, Clone, Copy, Hash)]
-pub struct DummyNotifier;
